% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_dirichlet.R
\name{aiken_classic}
\alias{aiken_classic}
\alias{aiken_V_classic}
\alias{aiken_H_classic}
\alias{aiken_C_classic}
\alias{aiken_R_classic}
\alias{aiken_A_classic}
\alias{aiken_I_classic}
\title{Coeficientes clasicos de Aiken}
\usage{
aiken_V_classic(x, k)

aiken_H_classic(x, k)

aiken_C_classic(x, k)

aiken_R_classic(x, y, k)

aiken_A_classic(focal, otros, k)

aiken_I_classic(focal, otros, k)
}
\arguments{
\item{x}{Vector de calificaciones enteras ya desplazadas al rango 0..k.}

\item{k}{Amplitud de la escala, \code{k = s - l = c - 1}.}

\item{y}{Segundo vector (solo para R).}

\item{focal}{Calificacion focal (solo para A e I).}

\item{otros}{Calificaciones de comparacion (solo para A e I).}
}
\value{
Un numero: el valor del coeficiente clasico.
}
\description{
Implementacion literal de las formulas originales, incluida la correccion de
paridad (\code{j = 0} si n es par, \code{j = 1} si n es impar) que suele
omitirse en las implementaciones disponibles.
}
