% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_table.R
\name{as.data.frame.aiken_table}
\alias{as.data.frame.aiken_table}
\title{Convertir un aiken_table a data.frame}
\usage{
\method{as.data.frame}{aiken_table}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Objeto \code{aiken_table}.}

\item{row.names, optional, ...}{Ignorados.}
}
\description{
Convertir un aiken_table a data.frame
}
