% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.bayes_coef}
\alias{as.data.frame.bayes_coef}
\title{Convertir bayes_coef a data.frame}
\usage{
\method{as.data.frame}{bayes_coef}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Objeto bayes_coef}

\item{row.names}{No usado}

\item{optional}{No usado}

\item{...}{Argumentos adicionales}
}
\description{
Convertir bayes_coef a data.frame
}
