% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{bayes_aiken_export}
\alias{bayes_aiken_export}
\title{Exportar resultados de bayes_aiken}
\usage{
bayes_aiken_export(x, file, format = "txt", include_verbose = TRUE)
}
\arguments{
\item{x}{Objeto bayes_aiken}

\item{file}{Ruta del archivo de salida}

\item{format}{Formato: "txt", "csv", "both" (default = "txt")}

\item{include_verbose}{Incluir salida verbose en txt (default = TRUE)}
}
\value{
Invisible NULL
}
\description{
Exporta los resultados a diferentes formatos
}
\examples{
ratings <- c(3, 3, 2, 3, 3, 2, 3, 3, 2, 3)
result <- bayes_aiken(ratings, coefficient = "V", l = 0, s = 3, verbose = FALSE)
# bayes_aiken_export(result, "output.txt")

}
