% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesAiken-package.R
\docType{package}
\name{BayesAiken-package}
\alias{BayesAiken}
\alias{BayesAiken-package}
\title{BayesAiken: Bayesian Content Validity Coefficients}
\description{
Bayesian inference for the six rating coefficients of Aiken (V, H, C, R,
A and I) with a Dirichlet-Multinomial model, together with the classical
coefficients, the Penfield-Giacobbi confidence interval for V and the
significance tests of Aiken (1985).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jventural/BayesAiken}
  \item Report bugs at \url{https://github.com/jventural/BayesAiken/issues}
}

}
\author{
\strong{Maintainer}: Jose Ventura-Leon \email{jventuraleon@gmail.com} (\href{https://orcid.org/0000-0003-2996-4244}{ORCID})

}
\keyword{internal}
