% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coef_A}
\alias{coef_A}
\title{Calcular Coeficiente A (Acuerdo del Juez) con intervalos bayesianos}
\usage{
coef_A(
  ratings_judge,
  ratings_others,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{ratings_judge}{Vector de calificaciones del juez focal}

\item{ratings_others}{Matriz de calificaciones de otros jueces}

\item{c}{Numero de categorias en la escala (default = 4)}

\item{prior_alpha}{Parametro alpha del prior Beta (default = 1)}

\item{prior_beta}{Parametro beta del prior Beta (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}

\item{verbose}{Mostrar resultados detallados (default = TRUE)}
}
\value{
Lista con resultados del coeficiente A
}
\description{
Calcular Coeficiente A (Acuerdo del Juez) con intervalos bayesianos
}
\examples{
judge <- c(2, 3, 3)
others <- matrix(c(2,3,2,2, 3,3,3,3, 3,3,3,3), nrow=3, byrow=TRUE)
result <- coef_A(judge, others, c = 4)

}
