% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_A_multi}
\alias{coef_A_multi}
\title{Calcular A para multiples jueces}
\usage{
coef_A_multi(
  ratings_matrix,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95
)
}
\arguments{
\item{ratings_matrix}{Matriz de calificaciones (filas = items, cols = jueces)}

\item{c}{Numero de categorias}

\item{prior_alpha}{Alpha del prior}

\item{prior_beta}{Beta del prior}

\item{cred_level}{Nivel de credibilidad}
}
\description{
Calcular A para multiples jueces
}
