% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coef_C}
\alias{coef_C}
\title{Calcular Coeficiente C (Consenso) con intervalos bayesianos}
\usage{
coef_C(
  ratings,
  l = 0,
  s = 3,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{ratings}{Vector de calificaciones de los jueces}

\item{l}{Valor minimo de la escala (default = 0)}

\item{s}{Valor maximo de la escala (default = 3)}

\item{prior_alpha}{Parametro alpha del prior Beta (default = 1)}

\item{prior_beta}{Parametro beta del prior Beta (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}

\item{verbose}{Mostrar resultados detallados (default = TRUE)}
}
\value{
Lista con resultados del coeficiente C
}
\description{
Calcular Coeficiente C (Consenso) con intervalos bayesianos
}
\examples{
ratings <- c(2, 2, 3, 2, 2)
result <- coef_C(ratings, l = 0, s = 3)

}
