% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coef_H}
\alias{coef_H}
\title{Calcular Coeficiente H (Homogeneidad) con intervalos bayesianos}
\usage{
coef_H(
  ratings,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{ratings}{Vector de calificaciones de los jueces}

\item{c}{Numero de categorias en la escala (default = 4)}

\item{prior_alpha}{Parametro alpha del prior Beta (default = 1)}

\item{prior_beta}{Parametro beta del prior Beta (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}

\item{verbose}{Mostrar resultados detallados (default = TRUE)}
}
\value{
Lista con resultados del coeficiente H
}
\description{
Calcular Coeficiente H (Homogeneidad) con intervalos bayesianos
}
\examples{
ratings <- c(2, 2, 3, 2, 2)
result <- coef_H(ratings, c = 4)

}
