% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_H_multi}
\alias{coef_H_multi}
\title{Calcular H para multiples items}
\usage{
coef_H_multi(
  data,
  item_col = NULL,
  criterion_col = NULL,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95
)
}
\arguments{
\item{data}{Data frame con calificaciones}

\item{item_col}{Nombre de columna de items}

\item{criterion_col}{Nombre de columna de criterios (opcional)}

\item{c}{Numero de categorias}

\item{prior_alpha}{Alpha del prior}

\item{prior_beta}{Beta del prior}

\item{cred_level}{Nivel de credibilidad}
}
\description{
Calcular H para multiples items
}
