% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coef_I}
\alias{coef_I}
\title{Calcular Coeficiente I (Consistencia Inter-Item) con intervalos bayesianos}
\usage{
coef_I(
  ratings_item1,
  ratings_item2,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{ratings_item1}{Vector de calificaciones del item 1}

\item{ratings_item2}{Vector de calificaciones del item 2}

\item{c}{Numero de categorias en la escala (default = 4)}

\item{prior_alpha}{Parametro alpha del prior Beta (default = 1)}

\item{prior_beta}{Parametro beta del prior Beta (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}

\item{verbose}{Mostrar resultados detallados (default = TRUE)}
}
\value{
Lista con resultados del coeficiente I
}
\description{
Calcular Coeficiente I (Consistencia Inter-Item) con intervalos bayesianos
}
\examples{
item1 <- c(2, 3, 3)
item2 <- c(3, 3, 3)
result <- coef_I(item1, item2, c = 4)

}
