% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coef_R}
\alias{coef_R}
\title{Calcular Coeficiente R (Reproducibilidad) con intervalos bayesianos}
\usage{
coef_R(
  ratings_t1,
  ratings_t2,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95,
  verbose = TRUE
)
}
\arguments{
\item{ratings_t1}{Vector de calificaciones en tiempo 1}

\item{ratings_t2}{Vector de calificaciones en tiempo 2}

\item{c}{Numero de categorias en la escala (default = 4)}

\item{prior_alpha}{Parametro alpha del prior Beta (default = 1)}

\item{prior_beta}{Parametro beta del prior Beta (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}

\item{verbose}{Mostrar resultados detallados (default = TRUE)}
}
\value{
Lista con resultados del coeficiente R
}
\description{
Calcular Coeficiente R (Reproducibilidad) con intervalos bayesianos
}
\examples{
t1 <- c(3, 3, 3, 3, 3)
t2 <- c(3, 3, 3, 2, 3)
result <- coef_R(t1, t2, c = 4)

}
