% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_R_multi}
\alias{coef_R_multi}
\title{Calcular R para multiples items (requiere datos en formato especial)}
\usage{
coef_R_multi(
  data_t1,
  data_t2,
  c = 4,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95
)
}
\arguments{
\item{data_t1}{Data frame con calificaciones tiempo 1}

\item{data_t2}{Data frame con calificaciones tiempo 2}

\item{c}{Numero de categorias}

\item{prior_alpha}{Alpha del prior}

\item{prior_beta}{Beta del prior}

\item{cred_level}{Nivel de credibilidad}
}
\description{
Calcular R para multiples items (requiere datos en formato especial)
}
