% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef_V_multi}
\alias{coef_V_multi}
\title{Calcular V de Aiken para multiples items}
\usage{
coef_V_multi(
  data,
  item_col = NULL,
  criterion_col = NULL,
  l = 0,
  s = 3,
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95
)
}
\arguments{
\item{data}{Data frame con calificaciones}

\item{item_col}{Nombre de columna de items}

\item{criterion_col}{Nombre de columna de criterios (opcional)}

\item{l}{Minimo de escala}

\item{s}{Maximo de escala}

\item{prior_alpha}{Alpha del prior}

\item{prior_beta}{Beta del prior}

\item{cred_level}{Nivel de credibilidad}
}
\value{
Data frame con resultados
}
\description{
Calcular V de Aiken para multiples items
}
