% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{compare_items}
\alias{compare_items}
\title{Comparar dos items usando diferencias posteriores}
\usage{
compare_items(result1, result2, delta = 0, n_samples = 10000, verbose = TRUE)
}
\arguments{
\item{result1}{Resultado del primer item (objeto de coef_V o lista)}

\item{result2}{Resultado del segundo item}

\item{delta}{Diferencia minima relevante (default = 0)}

\item{n_samples}{Numero de muestras Monte Carlo (default = 10000)}

\item{verbose}{Mostrar resultados detallados (default = TRUE)}
}
\value{
Lista con distribucion de diferencias y probabilidades
}
\description{
Calcula la distribucion posterior de la diferencia entre dos coeficientes
y reporta P(Delta > 0) o P(Delta > delta).
}
\examples{
ratings1 <- c(3, 3, 3, 3, 2, 3, 3, 3, 3, 3)
ratings2 <- c(2, 3, 2, 3, 2, 2, 3, 2, 3, 2)
res1 <- coef_V(ratings1, l = 0, s = 3, verbose = FALSE)
res2 <- coef_V(ratings2, l = 0, s = 3, verbose = FALSE)
compare_items(res1, res2)

}
