% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{compare_judge_groups}
\alias{compare_judge_groups}
\title{Comparar grupos de jueces}
\usage{
compare_judge_groups(
  ratings_group1,
  ratings_group2,
  l = 0,
  s = 3,
  delta = 0.05,
  prior_alpha = 1,
  prior_beta = 1,
  n_samples = 10000,
  verbose = TRUE
)
}
\arguments{
\item{ratings_group1}{Vector o matriz de calificaciones del grupo 1}

\item{ratings_group2}{Vector o matriz de calificaciones del grupo 2}

\item{l}{Valor minimo de la escala (default = 0)}

\item{s}{Valor maximo de la escala (default = 3)}

\item{delta}{Diferencia minima relevante para equivalencia (default = 0.05)}

\item{prior_alpha}{Alpha del prior (default = 1)}

\item{prior_beta}{Beta del prior (default = 1)}

\item{n_samples}{Numero de muestras Monte Carlo (default = 10000)}

\item{verbose}{Mostrar resultados (default = TRUE)}
}
\value{
Lista con comparacion de grupos
}
\description{
Compara las evaluaciones de dos grupos independientes de jueces
(ej. clinicos vs investigadores, Lima vs provincias).
}
\examples{
# Grupo 1: Clinicos
grupo1 <- c(3, 3, 2, 3, 3)
# Grupo 2: Investigadores
grupo2 <- c(3, 2, 3, 3, 2)
compare_judge_groups(grupo1, grupo2, l = 0, s = 3)

}
