% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_priors}
\alias{compare_priors}
\title{Comparar diferentes priors}
\usage{
compare_priors(
  ratings,
  l = 0,
  s = 3,
  priors = list(Uniform = c(1, 1), Jeffreys = c(0.5, 0.5), `Weakly Informative` = c(2,
    2), Skeptical = c(1, 3), Optimistic = c(3, 1)),
  cred_level = 0.95
)
}
\arguments{
\item{ratings}{Vector de calificaciones}

\item{l}{Valor minimo de la escala}

\item{s}{Valor maximo de la escala}

\item{priors}{Lista de priors a comparar}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}
}
\value{
Data frame con comparacion de priors
}
\description{
Compara los resultados usando diferentes distribuciones prior
}
\examples{
ratings <- c(3, 3, 2, 3, 3)
compare_priors(ratings, l = 0, s = 3)

}
