% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_hdi}
\alias{compute_hdi}
\title{Calcular HDI (Highest Density Interval)}
\usage{
compute_hdi(alpha, beta, cred_level = 0.95)
}
\arguments{
\item{alpha}{Parametro alpha de la distribucion Beta}

\item{beta}{Parametro beta de la distribucion Beta}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}
}
\value{
Vector con limites inferior y superior del HDI
}
\description{
Calcula el intervalo de mayor densidad para una distribucion Beta
}
\examples{
compute_hdi(25, 5, 0.95)

}
