% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_csv}
\alias{export_to_csv}
\title{Exportar a archivo CSV}
\usage{
export_to_csv(x, file)
}
\arguments{
\item{x}{Objeto bayes_aiken}

\item{file}{Ruta del archivo CSV}
}
\description{
Exportar a archivo CSV
}
