% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_txt}
\alias{export_to_txt}
\title{Exportar a archivo de texto}
\usage{
export_to_txt(x, file, include_verbose = TRUE)
}
\arguments{
\item{x}{Objeto bayes_aiken}

\item{file}{Ruta del archivo}

\item{include_verbose}{Incluir verbose (default = TRUE)}
}
\description{
Exportar a archivo de texto
}
