% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_results}
\alias{extract_results}
\title{Extraer resultados en formato data.frame}
\usage{
extract_results(..., ci_type = "HDI")
}
\arguments{
\item{...}{Objetos bayes_coef (de coef_V, coef_H, etc.) o bayes_aiken,
o una lista de objetos}

\item{ci_type}{Tipo de intervalo a incluir: "HDI" (default), "ETI", o "both"}
}
\value{
Data.frame con los resultados extraidos
}
\description{
Extrae los coeficientes y sus intervalos de credibilidad de uno o varios
objetos bayes_coef o bayes_aiken en un data.frame ordenado.
}
\examples{
# Un solo coeficiente
v1 <- coef_V(c(3,3,2,3,3), l=0, s=3, verbose=FALSE)
extract_results(v1)

# Multiples coeficientes
v2 <- coef_V(c(3,2,3,3,2), l=0, s=3, verbose=FALSE)
extract_results(v1, v2)

# Con nombres
extract_results(item1 = v1, item2 = v2)

}
