% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_dirichlet.R
\name{plot.aiken_bayes}
\alias{plot.aiken_bayes}
\title{Grafico de la posterior de un coeficiente de Aiken}
\usage{
\method{plot}{aiken_bayes}(x, ...)
}
\arguments{
\item{x}{Objeto \code{aiken_bayes}.}

\item{...}{Ignorado.}
}
\description{
Grafico de la posterior de un coeficiente de Aiken
}
