% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.bayes_aiken}
\alias{plot.bayes_aiken}
\title{=============================================================================
FUNCIONES DE VISUALIZACION - BayesAiken
=============================================================================
Plot del objeto bayes_aiken}
\usage{
\method{plot}{bayes_aiken}(x, show_prior = TRUE, show_hdi = TRUE, ...)
}
\arguments{
\item{x}{Objeto de clase bayes_aiken}

\item{show_prior}{Mostrar la distribucion prior (default = TRUE)}

\item{show_hdi}{Mostrar HDI (default = TRUE)}

\item{...}{Argumentos adicionales}
}
\description{
=============================================================================
FUNCIONES DE VISUALIZACION - BayesAiken
=============================================================================
Plot del objeto bayes_aiken
}
