% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{plot.bayes_comparison}
\alias{plot.bayes_comparison}
\title{Plot de comparacion de items}
\usage{
\method{plot}{bayes_comparison}(x, show_rope = FALSE, rope = c(-0.05, 0.05), ...)
}
\arguments{
\item{x}{Objeto de compare_items}

\item{show_rope}{Mostrar ROPE (default = FALSE)}

\item{rope}{Limites del ROPE si show_rope = TRUE}

\item{...}{No se usa.}
}
\description{
Plot de comparacion de items
}
