% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{plot.bayes_rope}
\alias{plot.bayes_rope}
\title{Plot del analisis ROPE}
\usage{
\method{plot}{bayes_rope}(x, ...)
}
\arguments{
\item{x}{Objeto bayes_rope}

\item{...}{Argumentos adicionales}
}
\description{
Plot del analisis ROPE
}
