% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{plot.bayes_sample_plan}
\alias{plot.bayes_sample_plan}
\title{Plot del planeamiento de muestra}
\usage{
\method{plot}{bayes_sample_plan}(x, ...)
}
\arguments{
\item{x}{Objeto bayes_sample_plan}

\item{...}{Argumentos adicionales}
}
\description{
Plot del planeamiento de muestra
}
