% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{plot.bayes_simulation}
\alias{plot.bayes_simulation}
\title{Grafico de resultados de simulacion}
\usage{
\method{plot}{bayes_simulation}(x, type = "all", ...)
}
\arguments{
\item{x}{Objeto bayes_simulation}

\item{type}{Tipo de grafico: "coverage" (cobertura), "width" (ancho),
"power" (Pr regla), o "all" (todos, default)}

\item{...}{Argumentos adicionales}
}
\description{
Grafico de resultados de simulacion
}
