% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_aiken_plots.R
\name{plot.bayes_V}
\alias{plot.bayes_V}
\title{=============================================================================
VISUALIZATION FUNCTIONS - BayesAiken Package
=============================================================================}
\usage{
\method{plot}{bayes_V}(x, show_prior = NULL, show_hdi = TRUE, ...)
}
\arguments{
\item{x}{Object of class 'bayes_V' from V_aiken()}

\item{show_prior}{Show prior distribution (default = TRUE for Beta, FALSE for Dirichlet)}

\item{show_hdi}{Show HDI shaded region (default = TRUE)}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot functions for Bayesian Aiken's V analysis.
}
\details{
=============================================================================
Plot Bayesian V Posterior Distribution
}
