% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.sensitivity_analysis}
\alias{plot.sensitivity_analysis}
\title{Grafico del analisis de sensibilidad}
\usage{
\method{plot}{sensitivity_analysis}(x, type = "all", ...)
}
\arguments{
\item{x}{Objeto sensitivity_analysis}

\item{type}{Tipo de grafico: "prior", "influence", "bootstrap", o "all" (default)}

\item{...}{Argumentos adicionales}
}
\description{
Grafico del analisis de sensibilidad
}
