% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all_coefficients}
\alias{plot_all_coefficients}
\title{Graficar todos los coeficientes}
\usage{
plot_all_coefficients(results, show_prior = FALSE)
}
\arguments{
\item{results}{Lista de resultados de coeficientes}

\item{show_prior}{Mostrar prior (default = FALSE)}
}
\description{
Graficar todos los coeficientes
}
