% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_comparison}
\alias{plot_comparison}
\title{Grafico de comparacion de items}
\usage{
plot_comparison(x, coef = NULL, order_by = "none")
}
\arguments{
\item{x}{Objeto bayes_aiken con multiples items}

\item{coef}{Coeficiente a graficar (si hay varios)}

\item{order_by}{Ordenar por: "none", "value", "prob" (default = "none")}
}
\description{
Grafico de comparacion de items
}
