% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_multiple_posteriors}
\alias{plot_multiple_posteriors}
\title{Graficar multiples items en un panel}
\usage{
plot_multiple_posteriors(
  results_list,
  n_cols = 4,
  show_prior = FALSE,
  main_title = "Posterior Distributions by Item",
  coef = "V",
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95
)
}
\arguments{
\item{results_list}{Lista de resultados o objeto bayes_aiken con multiples items}

\item{n_cols}{Numero de columnas en el panel (default = 4)}

\item{show_prior}{Mostrar prior (default = FALSE)}

\item{main_title}{Titulo principal del panel (opcional}

\item{coef}{Tipo de coeficiente (default = "V")}

\item{prior_alpha}{Alpha del prior (default = 1)}

\item{prior_beta}{Beta del prior (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}
}
\value{
Invisible NULL
}
\description{
Genera un panel con multiples graficos de distribucion posterior
}
\examples{
# Crear datos de ejemplo
items_data <- lapply(1:20, function(i) {
  list(post_alpha = 20 + sample(-5:10, 1),
       post_beta = 3 + sample(0:3, 1),
       item_name = paste0("Item ", i))
})
plot_multiple_posteriors(items_data, n_cols = 5)

}
