% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Graficar distribucion posterior}
\usage{
plot_posterior(
  post_alpha,
  post_beta,
  coef = "V",
  prior_alpha = 1,
  prior_beta = 1,
  cred_level = 0.95,
  show_prior = TRUE,
  show_hdi = TRUE,
  main = NULL,
  compact = FALSE
)
}
\arguments{
\item{post_alpha}{Parametro alpha de la posterior}

\item{post_beta}{Parametro beta de la posterior}

\item{coef}{Nombre del coeficiente}

\item{prior_alpha}{Parametro alpha del prior (default = 1)}

\item{prior_beta}{Parametro beta del prior (default = 1)}

\item{cred_level}{Nivel de credibilidad (default = 0.95)}

\item{show_prior}{Mostrar prior (default = TRUE)}

\item{show_hdi}{Mostrar HDI (default = TRUE)}

\item{main}{Titulo del grafico (opcional)}

\item{compact}{Modo compacto para paneles multiples (default = FALSE)}
}
\value{
Invisible NULL
}
\description{
Graficar distribucion posterior
}
\examples{
plot_posterior(post_alpha = 25, post_beta = 5, coef = "V")

}
