% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_dirichlet.R
\name{print.aiken_bayes}
\alias{print.aiken_bayes}
\title{Imprimir un objeto aiken_bayes}
\usage{
\method{print}{aiken_bayes}(x, ...)
}
\arguments{
\item{x}{Objeto de clase \code{aiken_bayes}.}

\item{...}{Ignorado.}
}
\description{
Imprimir un objeto aiken_bayes
}
