% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_table.R
\name{print.aiken_table}
\alias{print.aiken_table}
\title{Imprimir un objeto aiken_table}
\usage{
\method{print}{aiken_table}(x, ...)
}
\arguments{
\item{x}{Objeto \code{aiken_table}.}

\item{...}{Ignorado.}
}
\description{
Imprimir un objeto aiken_table
}
