% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.bayes_aiken}
\alias{print.bayes_aiken}
\title{Print method for bayes_aiken objects}
\usage{
\method{print}{bayes_aiken}(x, ...)
}
\arguments{
\item{x}{Objeto bayes_aiken}

\item{...}{Argumentos adicionales}
}
\description{
Print method for bayes_aiken objects
}
