% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.bayes_coef}
\alias{print.bayes_coef}
\title{Print method for bayes_coef objects}
\usage{
\method{print}{bayes_coef}(x, ...)
}
\arguments{
\item{x}{Objeto bayes_coef}

\item{...}{Argumentos adicionales}
}
\description{
Muestra el resultado de coef_V, coef_H, etc. en formato de tabla ordenada
}
