% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.R
\name{print.bayes_simulation}
\alias{print.bayes_simulation}
\title{Imprimir resultados de simulacion}
\usage{
\method{print}{bayes_simulation}(x, ...)
}
\arguments{
\item{x}{Objeto bayes_simulation}

\item{...}{Argumentos adicionales}
}
\description{
Imprimir resultados de simulacion
}
