% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.judge_heterogeneity}
\alias{print.judge_heterogeneity}
\title{Imprimir diagnostico de heterogeneidad}
\usage{
\method{print}{judge_heterogeneity}(x, ...)
}
\arguments{
\item{x}{Objeto judge_heterogeneity}

\item{...}{Argumentos adicionales}
}
\description{
Imprimir diagnostico de heterogeneidad
}
