% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.sensitivity_analysis}
\alias{print.sensitivity_analysis}
\title{Imprimir analisis de sensibilidad}
\usage{
\method{print}{sensitivity_analysis}(x, ...)
}
\arguments{
\item{x}{Objeto sensitivity_analysis}

\item{...}{Argumentos adicionales}
}
\description{
Imprimir analisis de sensibilidad
}
