% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_apa.R
\name{print_apa}
\alias{print_apa}
\title{Imprimir la tabla APA en consola}
\usage{
print_apa(x, ...)
}
\arguments{
\item{x}{Objeto \code{aiken_table} o data.frame de \code{aiken_apa}.}

\item{...}{Pasados a \code{\link{aiken_apa}}.}
}
\value{
Se invoca por su efecto.
}
\description{
Coloca cada celda en tres lineas alineadas, que es lo que hace legible el
formato. Un \code{data.frame} normal no lo consigue.
}
