% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{recommend_model}
\alias{recommend_model}
\title{Recomendar modelo para V de Aiken}
\usage{
recommend_model(
  ratings,
  l = 0,
  s = 3,
  threshold_n = 7,
  threshold_cv = 0.15,
  threshold_consensus = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{ratings}{Vector de calificaciones de los jueces}

\item{l}{Valor minimo de la escala (default = 0)}

\item{s}{Valor maximo de la escala (default = 3)}

\item{threshold_n}{Umbral de jueces para considerar muestra pequena (default = 7)}

\item{threshold_cv}{Umbral de CV para considerar alta heterogeneidad (default = 0.15)}

\item{threshold_consensus}{Umbral de proporcion en categoria modal (default = 0.80)}

\item{verbose}{Mostrar analisis detallado (default = TRUE)}
}
\value{
Lista con recomendacion y justificacion
}
\description{
Esta funcion analiza las caracteristicas de los datos y recomienda
cual modelo usar: Beta-Binomial (estandar) o Dirichlet-Multinomial
(conservador).
}
\details{
La funcion evalua tres criterios:
\itemize{
  \item Tamano muestral: n < threshold_n sugiere Dirichlet
  \item Heterogeneidad: CV > threshold_cv sugiere Dirichlet
  \item Consenso: proporcion modal < threshold_consensus sugiere Dirichlet
}

Regla de decision:
\itemize{
  \item 0 banderas rojas: Beta-Binomial
  \item 1 bandera roja: Beta-Binomial con precaucion
  \item 2+ banderas rojas: Dirichlet-Multinomial
}
}
\examples{
# Caso con alto consenso - recomienda Beta-Binomial
ratings_good <- c(3, 3, 3, 3, 3, 3, 3, 3, 2, 3)
recommend_model(ratings_good)

# Caso con baja n y heterogeneidad - recomienda Dirichlet
ratings_risky <- c(1, 2, 3, 2, 3)
recommend_model(ratings_risky)

}
