% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiken_dirichlet.R
\name{summary.aiken_bayes}
\alias{summary.aiken_bayes}
\title{Resumen de un objeto aiken_bayes}
\usage{
\method{summary}{aiken_bayes}(object, ...)
}
\arguments{
\item{object}{Objeto \code{aiken_bayes}.}

\item{...}{Ignorado.}
}
\description{
Resumen de un objeto aiken_bayes
}
