% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.bayes_aiken}
\alias{summary.bayes_aiken}
\title{Summary method for bayes_aiken objects}
\usage{
\method{summary}{bayes_aiken}(object, ...)
}
\arguments{
\item{object}{Objeto bayes_aiken}

\item{...}{Argumentos adicionales}
}
\description{
Summary method for bayes_aiken objects
}
