% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{summary_table}
\alias{summary_table}
\title{Crear tabla resumen formateada}
\usage{
summary_table(x, digits = 3)
}
\arguments{
\item{x}{Objeto bayes_aiken}

\item{digits}{Numero de decimales (default = 3)}
}
\value{
Data frame con resumen formateado
}
\description{
Crear tabla resumen formateada
}
